Change responsibility of anyone oppressed
I originally asked anyone who had been oppressed to contribute by taking on responsibility of educating and informing, but this was phrased too much like a demand and without any acknowledgement of cost. To being to address this, I have rephrased and made an attempt to offer recompensation in kind (e.g. by contributions to the contributor's own projects).